requirements.txt
py-solc-x~=1.1.1
web3~=6.0.0b3
